reworking API routes
